Refactor Code to modularize utils
