+++ blueprints/docs.py
@@ -1,5 +1,4 @@
-from flask import Blueprint
-from utils import *
+from flask import Blueprint, render_template, request
 from .limiter import limiter
 import os
