+++ blueprints/url_shortener.py
@@ -1,10 +1,47 @@
 import time
-from flask import Blueprint
-from utils import *
+from flask import (
+    Blueprint,
+    request,
+    jsonify,
+    render_template,
+    redirect,
+    url_for,
+    make_response,
+)
+from utils.url_utils import (
+    BOT_USER_AGENTS,
+    get_country,
+    get_client_ip,
+    validate_password,
+    validate_url,
+    validate_alias,
+    validate_emoji_alias,
+    validate_expiration_time,
+    generate_short_code,
+    generate_emoji_alias,
+    convert_to_gmt,
+)
+from utils.mongo_utils import (
+    load_url,
+    update_url,
+    insert_url,
+    load_emoji_url,
+    update_emoji_url,
+    insert_emoji_url,
+    check_if_slug_exists,
+    check_if_emoji_alias_exists,
+    validate_blocked_url,
+    urls_collection,
+)
+from utils.general import is_positive_integer, humanize_number
 from .limiter import limiter
 from .cache import cache
 
 from user_agents import parse
+import json
+from datetime import datetime, timezone
+from urllib.parse import unquote
+import re
 import tldextract
 from crawlerdetect import CrawlerDetect
 
@@ -60,7 +97,7 @@ def shorten_url():
     if url and not validate_blocked_url(url):
         return jsonify({"BlockedUrlError": "Blocked URL ⛔"}), 403
 
-    if alias and not validate_string(alias):
+    if alias and not validate_alias(alias):
         if request.headers.get("Accept") == "application/json":
             return jsonify({"AliasError": "Invalid Alias", "alias": f"{alias}"}), 400
         else:
@@ -150,7 +187,7 @@ def shorten_url():
 
     data["creation-ip-address"] = get_client_ip()
 
-    add_url_by_id(short_code, data)
+    insert_url(short_code, data)
 
     response = jsonify({"short_url": f"{request.host_url}{short_code}"})
 
@@ -259,7 +296,7 @@ def emoji():
 
     data["creation-ip-address"] = get_client_ip()
 
-    add_emoji_by_alias(emojies, data)
+    insert_emoji_url(emojies, data)
 
     response = jsonify({"short_url": f"{request.host_url}{emojies}"})
 
@@ -275,9 +312,9 @@ def result(short_code):
 
     short_code = unquote(short_code)
     if validate_emoji_alias(short_code):
-        url_data = load_emoji_by_alias(short_code)
+        url_data = load_emoji_url(short_code)
     else:
-        url_data = load_url_by_id(short_code)
+        url_data = load_url(short_code)
 
     if url_data:
         short_code = url_data["_id"]
@@ -326,9 +363,9 @@ def redirect_url(short_code):
 
     if validate_emoji_alias(short_code):
         is_emoji = True
-        url_data = db["emojis"].find_one({"_id": short_code}, projection)
+        url_data = load_emoji_url(short_code, projection)
     else:
-        url_data = db["urls"].find_one({"_id": short_code}, projection)
+        url_data = load_url(short_code, projection)
 
     if not url_data:
         return (
@@ -387,7 +424,7 @@ def redirect_url(short_code):
 
     try:
         ua = parse(user_agent)
-    except TypeError as e:
+    except TypeError:
         return "Invalid User-Agent", 400
 
     os_name = ua.os.family
@@ -471,7 +508,7 @@ def redirect_url(short_code):
     updates["$inc"][f"counter.{today}"] = 1
 
     if "ips" in url_data:
-        if not user_ip in url_data["ips"]:
+        if user_ip not in url_data["ips"]:
             updates["$inc"][f"unique_counter.{today}"] = 1
     else:
         updates["$inc"][f"unique_counter.{today}"] = 1
@@ -501,9 +538,9 @@ def redirect_url(short_code):
     )
 
     if is_emoji:
-        db["emojis"].update_one({"_id": short_code}, updates)
+        update_emoji_url(short_code, updates)
     else:
-        db["urls"].update_one({"_id": short_code}, updates)
+        update_url(short_code, updates)
 
     return redirect(url)
 
@@ -519,9 +556,9 @@ def check_password(short_code):
 
     short_code = unquote(short_code)
     if validate_emoji_alias(short_code):
-        url_data = db["emojis"].find_one({"_id": short_code}, projection)
+        url_data = load_emoji_url(short_code, projection)
     else:
-        url_data = db["urls"].find_one({"_id": short_code}, projection)
+        url_data = load_url(short_code, projection)
 
     if url_data:
         # check if the URL is password protected
@@ -564,7 +601,7 @@ METRIC_PIPELINE = [
 @limiter.exempt
 @cache.cached(timeout=60)
 def metric():
-    result = collection.aggregate(METRIC_PIPELINE).next()
+    result = urls_collection.aggregate(METRIC_PIPELINE).next()
     del result["_id"]
     result["total-clicks"] = humanize_number(result["total-clicks"])
     result["total-shortlinks"] = humanize_number(result["total-shortlinks"])
